requirements.txt
python-telegram-bot==13.15
pyTelegramBotAPI
pytesseract
Pillow
openai
python-dotenv